Add scheduler for redis tasks
