+++ app/__init__.py
@@ -1,11 +1,11 @@
 from flask import Flask
 from config import Config
 from app.models import db
-from app.cache_report import redis_client
-from app import models
+from app import models, db
 from flask_migrate import Migrate
 from flask_login import LoginManager
 from flask_bootstrap import Bootstrap
+from flask_redis import FlaskRedis
 from app.auth.routes import auth
 from app.main.routes import main
 from app.report.routes import report
@@ -14,13 +14,14 @@ from app.api.endpoints import api_bp
 from flask_jwt_extended import JWTManager
 from redis import Redis
 import rq
-
+from rq_scheduler import Scheduler
 
 migrate = Migrate()
 login = LoginManager()
 login.login_view = 'auth.login'
 bootstrap = Bootstrap()
 jwt = JWTManager()
+redis_client = FlaskRedis()
 
 
 def create_app(config_class=Config):
@@ -36,6 +37,7 @@ def create_app(config_class=Config):
     jwt.init_app(app)
     app.redis = Redis.from_url(app.config['REDIS_URL'])
     app.task_queue = rq.Queue('affiliate', connection=app.redis)
+    app.scheduler = Scheduler(queue=app.task_queue, connection=app.redis)
 
     app.register_blueprint(auth)
     app.register_blueprint(main)
