+++ app/report/routes.py
@@ -9,18 +9,21 @@ report = Blueprint('report', __name__)
 @report.route('/report/all_actions_report')
 @login_required
 def all_actions_report():
+    current_user.user_caching_report()
+
     page = request.args.get('page', 1, type=int)
     actions = (
         Action.query
         .join(Link)
         .filter_by(user_id=current_user.id)
         .order_by(Action.timestamp.desc())
+        .paginate(page=page, per_page=25)
     )
-    page_actions = actions.paginate(page=page, per_page=25)
-    prev_page = url_for('report.all_actions_report', page=page_actions.prev_num) if page_actions.has_prev else None
-    next_page = url_for('report.all_actions_report', page=page_actions.next_num) if page_actions.has_next else None
+
+    prev_page = url_for('report.all_actions_report', page=actions.prev_num) if actions.has_prev else None
+    next_page = url_for('report.all_actions_report', page=actions.next_num) if actions.has_next else None
     return render_template('reports/all_actions_report.html',
-                           actions=page_actions.items,
+                           actions=actions.items,
                            prev=prev_page,
                            next=next_page,
                            )
